move similarity code into script
