+++ chemprop/data/__init__.py
@@ -1,4 +1,3 @@
 from .data import MoleculeDatapoint, MoleculeDataset
-from .scaffold import cluster_split
+from .scaffold import cluster_split, scaffold_to_smiles
 from .scaler import StandardScaler
-from .similarity import morgan_similarity, scaffold_similarity
